Renamed Token to InviteToken Secured MongoDB
