+++ config/custom-environment-variables.json
@@ -8,5 +8,6 @@
     "beerpong-sign-key": "BEERPONG_SIGN_KEY",
     "beerpong-email-smtp-host": "BEERPONG_EMAIL_SMTP_HOST",
     "beerpong-email-smtp-port": "BEERPONG_EMAIL_SMTP_PORT",
-    "beerpong-invite-url": "BEERPONG_INVITE_URL"
+    "beerpong-invite-url": "BEERPONG_INVITE_URL",
+    "beerpong-mongo-initdb-root-password": "BEERPONG_MONGO_INITDB_ROOT_PASSWORD"
 }
