+++ docker-compose.yml
@@ -24,6 +24,7 @@ services:
       - BEERPONG_EMAIL_SMTP_PORT=${BEERPONG_EMAIL_SMTP_PORT}
       - BEERPONG_SIGN_KEY=${BEERPONG_SIGN_KEY}
       - BEERPONG_INVITE_URL=${BEERPONG_INVITE_URL}
+      - BEERPONG_MONGO_INITDB_ROOT_PASSWORD=${BEERPONG_MONGO_INITDB_ROOT_PASSWORD}
   mongo:
     container_name: mongo
     image: mongo
@@ -31,3 +32,6 @@ services:
       - "27017:27017"
     volumes:
       - /beerpong-be/beerpong-be-mongo-data:/data/db
+    environment:
+      - MONGO_INITDB_ROOT_USERNAME=root
+      - MONGO_INITDB_ROOT_PASSWORD=${BEERPONG_MONGO_INITDB_ROOT_PASSWORD}
